UI changes

+++ app/src/main/java/bidding/example/com/bidding/Lottery_Game/Multiple_Bet.java
@@ -651,7 +651,7 @@ public class Multiple_Bet extends Fragment implements View.OnClickListener{
                     if (connectionDetector.isConnectingToInternet()) {
                         for (int i = 0; i < 10; i++) {
                             if (amount2[i] != null) {
-                                if(Integer.parseInt(amount2[i])>Double.parseDouble(result)){
+                                if(Double.parseDouble(amount2[i])>Double.parseDouble(result)){
                                     Toast.makeText(getActivity(), "Not enough chips found", Toast.LENGTH_SHORT).show();
                                     return;
                                 }
@@ -885,7 +885,7 @@ public class Multiple_Bet extends Fragment implements View.OnClickListener{
                         {
                             if(amount3[i] != null)
                             {
-                                if(Integer.parseInt(amount3[i])>Double.parseDouble(result)){
+                                if(Double.parseDouble(amount3[i])>Double.parseDouble(result)){
                                     Toast.makeText(getActivity(), "Not enough chips found", Toast.LENGTH_SHORT).show();
                                     return;
                                 }
